Converted "medicate delay" time to use units days in code, and similarly for drug inputs.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@709 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/PkPd/HoshenPkPdModel.h
@@ -47,7 +47,7 @@ public:
   virtual void checkpoint (ostream& stream);
   
   //TODO: do we need to pass age?
-  virtual void medicate(string drugAbbrev, double qty, int time, double age);
+  virtual void medicate(string drugAbbrev, double qty, double time, double age);
   virtual void decayDrugs (double ageYears);
   virtual double getDrugFactor (uint32_t proteome_ID, double ageYears);
