+++ include/PkPd/LSTMPkPdModel.h
@@ -46,7 +46,7 @@ public:
   virtual void checkpoint (istream& stream);
   virtual void checkpoint (ostream& stream);
   
-  virtual void medicate(string drugAbbrev, double qty, int time, double age);
+  virtual void medicate(string drugAbbrev, double qty, double time, double age);
   virtual void decayDrugs (double ageYears);
   virtual double getDrugFactor (uint32_t proteome_ID, double ageYears);
