+++ include/WithinHost/Empirical.h
@@ -44,7 +44,7 @@ public:
     virtual void newInfection();
     virtual void clearAllInfections();
     
-    virtual void medicate (string drugName, double qty, int time, double age);
+    virtual void medicate (string drugName, double qty, double time, double age);
     
     /** Update densities for timestep (taking into account blood-stage vaccine
      * and drug efficacies. */
