+++ model/WithinHost/Dummy.cpp
@@ -56,7 +56,7 @@ void DummyWithinHostModel::clearAllInfections(){
 
 // -----  medicate drugs -----
 
-void DummyWithinHostModel::medicate(string drugName, double qty, int time, double age) {
+void DummyWithinHostModel::medicate(string drugName, double qty, double time, double age) {
   pkpdModel->medicate(drugName, qty, time, age);
 }
